# docs/swagger.yaml
definitions:
  common.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  common.ResponseSocialMedia:
    properties:
      data:
        $ref: '#/definitions/common.SocialMediaDTO'
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  common.SocialMediaDTO:
    properties:
      social_medias: {}
    type: object
...
  model.SocialMediaCreateRequestSwaggo:
    properties:
      name:
        example: Instagram
        type: string
      social_media_url:
        example: www.instagram.com
        type: string
    type: object
  model.SocialMediaUpdateRequestSwaggo:
    properties:
      name:
        example: Instagram
        type: string
      social_media_url:
        example: www.instagram.com
        type: string
    type: object
  model.UserLoginRequestSwaggo:
    properties:
      email:
        example: viky@gmail.com
        type: string
      password:
        example: password
        type: string
    type: object
  model.UserRegisterRequestSwaggo:
    properties:
      age:
        example: 20
        type: integer
      email:
        example: viky@gmail.com
        type: string
      password:
        example: password
        type: string
      username:
        example: viky
        type: string
    type: object
  model.UserUpdateRequestSwaggo:
    properties:
      email:
        example: xxx@gmail.com
        type: string
      username:
        example: xxx
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: User Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      summary: User Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User Register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRegisterRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      summary: User Register
      tags:
      - Auth
...
  /social-media:
    get:
      consumes:
      - application/json
      description: Find all social media
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.ResponseSocialMedia'
            type: array
      security:
      - Bearer: []
      summary: Find all social media
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Create a new social media
      parameters:
      - description: Social Media Create Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.SocialMediaCreateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Create a new social media
      tags:
      - Social Media
  /social-media/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Delete a social media
      parameters:
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Delete a social media
      tags:
      - Social Media
    get:
      consumes:
      - application/json
      description: Find one social media
      parameters:
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Find one social media
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update one social media
      parameters:
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: string
      - description: Social Media Update Request
        in: body
        name: socialMedia
        required: true
        schema:
          $ref: '#/definitions/model.SocialMediaUpdateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Update one social media
      tags:
      - Social Media
  /users:
    delete:
      consumes:
      - application/json
      description: Delete a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: User Delete
      tags:
      - Users
  /users/{userId}:
    put:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: User Update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: User Update
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"